Backend is fixed

+++ orchestrating-docker/web/app.py
@@ -69,7 +69,7 @@ post_enterquestion_model = api.model('Question', {'Question': fields.String,
 get_enrollment_model = api.model('enrollment', {'course_number': fields.String})
 post_enrollment_model = api.model('Enrollment', {'course_number' : fields.String})
 get_iclickerresponse_model = api.model('get_iclicker_response', {'iqid': fields.Integer(description = 'Question number to get'), 'sessionid':fields.String})
-post_iclickerresponse_model = api.model('iClicker_post_responses', {'response': fields.String})
+post_iclickerresponse_model = api.model('iClicker_post_responses', {'response': fields.Integer})
 login_model = api.model('login', {
     'netid': fields.String(description='netid', required=True),
     'password': fields.String(description='Password', required=True)
@@ -80,7 +80,7 @@ create_user_model = api.model('create_user', {
     'lastname': fields.String(description='lastname', required=True),
     'password': fields.String(description='Password', required=True)
 })
-put_upvotes_model = api.model('upvotes',{})
+#put_upvotes_model = api.model('upvotes', {})
 get_iclickerquestion_model = api.model('iclicker_question_get',{})
 post_iclickerquestion_model = api.model('iclicker_question_post', {'sessionid': fields.String,
                                                                    'question': fields.String,
@@ -286,9 +286,11 @@ class sessioninformation(Resource):
                     response = db.engine.execute(sessionInfo, sessionid=sessionid).fetchall()
                     updatets = text('UPDATE session SET readts = :ts WHERE sessionid=:sessionid')
                     db.engine.execute(updatets, ts=ts, sessionid=sessionid)
+                    endTransaction()
                 else:
                     response = "[]"
-                endTransaction()
+                    endTransaction()
+                    return response
             except psycopg2.Error:
                 rollBack()
             else:
@@ -314,7 +316,12 @@ class sessioninformation(Resource):
                 updatets = text('UPDATE session SET readts = :ts WHERE sessionid = :sessionid')
                 db.engine.execute(updatets, ts=ts, sessionid=sessionid)
 
-                if sameSession is None:
+                checkCourseNumber = text('SELECT * FROM courses WHERE course_number=:course_number')
+                courseExists = db.engine.execute(checkCourseNumber, course_number=course_number).fetchone()
+                updatets = text('UPDATE courses SET readts = :ts WHERE course_number = :course_number')
+                db.engine.execute(updatets, ts=ts, course_number=course_number)
+
+                if sameSession is None and courseExists is not None:
                     newSession = text('INSERT INTO session (sessionid, date, term, course_number, writets) VALUES (:sessionid, :date, :term, :course_number, :ts)')
                     db.engine.execute(newSession, sessionid=sessionid, date=date_key, term=term, course_number=course_number, ts=ts)
                 #pass this sessionid to every otehr table
@@ -338,49 +345,53 @@ class sessioninformation(Resource):
                 date_key = str(datetime.datetime.now().month)+"-"+str(datetime.datetime.now().day)
                 hash_key = date_key+term+course_number
                 sessionid = hashlib.sha256(hash_key.encode('utf-8')).hexdigest()
+                checkSessionid = text('SELECT * FROM session WHERE sessionid = :sessionid')
+                exists = db.engine.execute(checkSessionid, sessionid=sessionid).fetchone()
 
-                # post to piazza, then delete questions
-                questionInfo = text('SELECT ques FROM student_question WHERE sessionid = :sessionid')
-                questions = db.engine.execute(questionInfo, sessionid=sessionid).fetchall()
-                updatets = text('UPDATE student_question SET readts = :ts WHERE sessionid = :sessionid')
-                db.engine.execute(updatets, ts=ts, sessionid=sessionid)
 
-                # get networkid, netid and password of a student
-                courseInfo = text('SELECT piazza_nid, piazza_netid, piazza_passwd FROM courses WHERE course_number=:course_number AND term=:term')
-                piazzaTuple = db.engine.execute(courseInfo, course_number=course_number, term=term).fetchone()
-                if piazzaTuple is not None:
-                    piazza_nid, piazza_netid, piazza_passwd = piazzaTuple[0], piazzaTuple[1], piazzaTuple[2]
-                    updatets = text('UPDATE courses SET readts = :ts WHERE course_number = :course_number AND term = :term')
-                    db.engine.execute(updatets, ts=ts, course_number=course_number, term=term)
+                if exists is not None:
+                    # post to piazza, then delete questions
+                    questionInfo = text('SELECT ques FROM student_question WHERE sessionid = :sessionid')
+                    questions = db.engine.execute(questionInfo, sessionid=sessionid).fetchall()
+                    updatets = text('UPDATE student_question SET readts = :ts WHERE sessionid = :sessionid')
+                    db.engine.execute(updatets, ts=ts, sessionid=sessionid)
 
-                    piazzaMigration(questions, piazza_nid, piazza_netid, piazza_passwd)
+                    # get networkid, netid and password of a student
+                    courseInfo = text('SELECT piazza_nid, piazza_netid, piazza_passwd FROM courses WHERE course_number=:course_number AND term=:term')
+                    piazzaTuple = db.engine.execute(courseInfo, course_number=course_number, term=term).fetchone()
+                    if piazzaTuple is not None:
+                        piazza_nid, piazza_netid, piazza_passwd = piazzaTuple[0], piazzaTuple[1], piazzaTuple[2]
+                        updatets = text('UPDATE courses SET readts = :ts WHERE course_number = :course_number AND term = :term')
+                        db.engine.execute(updatets, ts=ts, course_number=course_number, term=term)
 
-                # Purge everything
-                purgeQuestions = text('DELETE FROM student_question WHERE sessionid = :sessionid')
-                db.engine.execute(purgeQuestions, sessionid=sessionid)
-                purgeUpvotes = text('DELETE FROM upvotes WHERE sessionid = :sessionid')
-                db.engine.execute(purgeUpvotes, sessionid=sessionid)
+                        piazzaMigration(questions, piazza_nid, piazza_netid, piazza_passwd)
 
-                #Grade responses for IClicker questions
-                total_questions_query = text('SELECT Count(*) FROM iclickerquestion WHERE sessionid = :sessionid')
-                total_questions = db.engine.execute(total_questions_query, sessionid=sessionid).scalar()
-                total_students_query = text('SELECT netid, Count(iqid) AS questions_answered FROM iclickerresponse, iclickerquestion WHERE sessionid = :sessionid AND iclickerquestion.sessionid = iclickerresponse.sessionid AND iclickerquestion.iqid = iclickerresponse.iqid AND iclickerresponse.responsetime BETWEEN iclickerquestion.starttime AND iclickerquestion.endtime GROUP BY netid')
-                total_students = db.engine.execute(total_students_query, sessionid=sessionid).fetchall()
-                for student in total_students:
-                    netid = student[0]
-                    answers = student[1]
-                    grade = answers/total_questions
-                    enrollment_grade_query = text('UPDATE enrollment SET grade=:grade, writets=:ts WHERE course_number=:course_number AND term=:term AND netid=:netid').fetchone()
-                    db.engine.execute(enrollment_grade_query, grade=grade, course_number=course_number, term=term, netid=netid, ts=ts)
+                    # Purge everything
+                    purgeQuestions = text('DELETE FROM student_question WHERE sessionid = :sessionid')
+                    db.engine.execute(purgeQuestions, sessionid=sessionid)
+                    purgeUpvotes = text('DELETE FROM upvotes WHERE sessionid = :sessionid')
+                    db.engine.execute(purgeUpvotes, sessionid=sessionid)
 
-                purgeIClickerQuestions = text('DELETE FROM iclickerquestion WHERE sessionid = :sessionid')
-                db.engine.execute(purgeIClickerQuestions, sessionid=sessionid)
-                purgeIClickerResponses = text('DELETE FROM iclickerresponse WHERE sessionid = :sessionid')
-                db.engine.execute(purgeIClickerResponses, sessionid=sessionid)
+                    #Grade responses for IClicker questions
+                    total_questions_query = text('SELECT Count(*) FROM iclickerquestion WHERE sessionid = :sessionid')
+                    total_questions = db.engine.execute(total_questions_query, sessionid=sessionid).scalar()
+                    total_students_query = text('SELECT netid, Count(iclickerresponse.iqid) AS questions_answered FROM iclickerresponse, iclickerquestion WHERE iclickerresponse.sessionid = :sessionid AND iclickerquestion.sessionid = iclickerresponse.sessionid AND iclickerquestion.iqid = iclickerresponse.iqid AND iclickerresponse.responsetime BETWEEN iclickerquestion.starttime AND iclickerquestion.endtime GROUP BY netid')
+                    total_students = db.engine.execute(total_students_query, sessionid=sessionid).fetchall()
+                    for student in total_students:
+                        netid = student[0]
+                        answers = student[1]
+                        answers = float(answers)
+                        grade = answers/total_questions
+                        enrollment_grade_query = text('UPDATE enrollment SET grade=:grade, writets=:ts WHERE course_number=:course_number AND term=:term AND netid=:netid')
+                        db.engine.execute(enrollment_grade_query, grade=grade, course_number=course_number, term=term, netid=netid, ts=ts)
 
-                purgeSession = text('DELETE FROM session WHERE sessionid=:sessionid')
-                db.engine.execute(purgeSession, sessionid=sessionid)
+                    purgeIClickerQuestions = text('DELETE FROM iclickerquestion WHERE sessionid = :sessionid')
+                    db.engine.execute(purgeIClickerQuestions, sessionid=sessionid)
+                    purgeIClickerResponses = text('DELETE FROM iclickerresponse WHERE sessionid = :sessionid')
+                    db.engine.execute(purgeIClickerResponses, sessionid=sessionid)
 
+                    purgeSession = text('DELETE FROM session WHERE sessionid=:sessionid')
+                    db.engine.execute(purgeSession, sessionid=sessionid)
                 endTransaction()
             except psycopg2.Error:
                 rollBack()
@@ -400,20 +411,21 @@ class Iclickerquestion(Resource):
                 ts = startTransaction()
                 sessionid = get_sessionid_student(netid,ts)
                 if sessionid is not None:
-                    print("session id not none")
-                    questionInfo = text('SELECT ques FROM iclickerquestion WHERE sessionid = :sessionid')
+                    #print("session id not none")
+                    questionInfo = text('SELECT ques, optiona, optionb, optionc, optiond FROM iclickerquestion WHERE sessionid = :sessionid')
                     response = db.engine.execute(questionInfo, sessionid=sessionid).fetchall()
                     updatets = text('UPDATE iclickerquestion SET readts = :ts WHERE sessionid = :sessionid')
                     db.engine.execute(updatets, ts=ts, sessionid=sessionid)
+                    endTransaction()
                 else:
-                    print("session id is none")
+                    #print("session id is none")
                     response = "[]"
-                endTransaction()
+                    endTransaction()
+                    return response
             except psycopg2.Error:
                 rollBack()
             else:
                 break
-        #return jsonify({'response' : [dict(row) for row in response]})
         return json.dumps([dict(row) for row in response])
 
     @api.expect(post_iclickerquestion_model)
@@ -474,6 +486,10 @@ class StudentEnrollment(Resource):
                 response = db.engine.execute(enrollmentInfo, netid=netid).fetchall()
                 updatets = text('UPDATE enrollment SET readts = :ts WHERE netid=:netid')
                 db.engine.execute(updatets, ts=ts, netid=netid)
+                if response is None:
+                    response = "[]"
+                    endTransaction()
+                    return response
                 endTransaction()
             except psycopg2.Error:
                 rollBack()
@@ -498,12 +514,17 @@ class StudentEnrollment(Resource):
                 updatets = text('UPDATE enrollment SET readts = :ts WHERE netid=:netid AND course_number=:course_number AND term=:term')
                 db.engine.execute(updatets, ts=ts, netid=netid, course_number=course_number, term=term)
 
-                if alreadyEnrolled is None:
-                    newEnrollment = text('INSERT INTO enrollment (netid, course_number, term, writets) VALUES (:netid, :course_number, :term, :ts)')
+                checkCourse = text('SELECT * FROM courses WHERE course_number=:course_number AND term=:term')
+                courseExists = db.engine.execute(checkCourse, course_number=course_number, term=term).fetchone()
+                updatets = text('UPDATE courses SET readts = :ts WHERE course_number=:course_number AND term=:term')
+                db.engine.execute(updatets, ts=ts, course_number=course_number, term=term)
+
+                if alreadyEnrolled is None and courseExists is not None:
+                    newEnrollment = text('INSERT INTO enrollment (netid, course_number, term, grade, writets) VALUES (:netid, :course_number, :term, 0, :ts)')
                     db.engine.execute(newEnrollment, netid=netid, course_number=course_number, term=term, ts=ts)
-                # Sets piazza netid for this course only if it is null
-                updatePiazzaNetid = text('UPDATE courses SET piazza_netid=:netid, writets=:ts WHERE course_number=:course_number AND term=:term AND piazza_netid IS NULL')
-                db.engine.execute(updatePiazzaNetid, netid=netid, ts=ts, course_number=course_number, term=term)
+                    # Sets piazza netid for this course only if it is null
+                    updatePiazzaNetid = text('UPDATE courses SET piazza_netid=:netid, writets=:ts WHERE course_number=:course_number AND term=:term AND piazza_netid IS NULL')
+                    db.engine.execute(updatePiazzaNetid, netid=netid, ts=ts, course_number=course_number, term=term)
                 endTransaction()
             except psycopg2.Error:
                 rollBack()
@@ -526,9 +547,11 @@ class Iclickerresponseget(Resource):
                     response = db.engine.execute(iclickerresponseInfo, sessionid=sessionid, iqid=iqid).fetchall()
                     updatets = text('UPDATE iclickerresponse SET readts = :ts WHERE sessionid = :sessionid AND iqid = :iqid')
                     db.engine.execute(updatets, ts=ts, sessionid=sessionid, iqid=iqid)
+                    endTransaction()
                 else:
                     response = "[]"
-                endTransaction()
+                    endTransaction()
+                    return response
             except psycopg2.Error:
                 rollBack()
             else:
@@ -560,7 +583,7 @@ class Iclickerresponseput(Resource):
                         iqid = 1
                     else:
                         get_latest_qid = text('SELECT MAX(iqid) FROM iclickerresponse WHERE sessionid = :sessionid')
-                        latest_qid = db.engine.execute(latest_qid, sessionid=sessionid).scalar()
+                        latest_qid = db.engine.execute(get_latest_qid, sessionid=sessionid).scalar()
                         #update the timestamp
                         updatets = text('UPDATE iclickerresponse SET readts = :ts WHERE sessionid = :sessionid')
                         db.engine.execute(updatets, ts=ts, sessionid=sessionid)
@@ -592,9 +615,11 @@ class StudentQuestionPost(Resource):
                     response = db.engine.execute(questionInfo, sessionid=sessionid).fetchall()
                     updatets = text('UPDATE student_question SET readts = :ts WHERE sessionid = :sessionid')
                     db.engine.execute(updatets, ts=ts, sessionid=sessionid)
+                    endTransaction()
                 else:
                     response = "[]"
-                endTransaction()
+                    endTransaction()
+                    return response
             except psycopg2.Error:
                 rollBack()
             else:
@@ -643,8 +668,7 @@ class StudentQuestionPost(Resource):
 
 @q_api.route('/<qid>')
 class UpvotesPost(Resource):
-    @api.expect(put_upvotes_model)
-    @api.doc(body=put_upvotes_model)
+    #@api.doc(body=put_upvotes_model)
     @jwt_required
     def put(self, qid):
         netid = get_jwt_identity()
@@ -694,6 +718,10 @@ class Courses(Resource):
                 response = db.engine.execute(courseInfo, term=term).fetchall()
                 updatets = text('UPDATE courses SET readts = :ts WHERE term=:term')
                 db.engine.execute(updatets, ts=ts, term=term)
+                if response is None:
+                    response = "[]"
+                    endTransaction()
+                    return response
                 endTransaction()
             except psycopg2.Error:
                 rollBack()
